Changed some item prices to two bytes each.

+++ OpenTESArena/src/Assets/ExeData.h
@@ -237,7 +237,7 @@ public:
 		std::array<std::string, 14> armorEnchantmentNames;
 		std::array<uint8_t, 14> armorEnchantmentQualities;
 		std::array<uint8_t, 14> armorEnchantmentSpells; // 255 means attribute bonus instead of spell.
-		std::array<uint8_t, 14> armorEnchantmentBonusPrices;
+		std::array<uint16_t, 14> armorEnchantmentBonusPrices;
 
 		// Weapon values.
 		std::array<std::string, 18> weaponNames; // Staff, ..., long bow.
@@ -251,13 +251,13 @@ public:
 		std::array<std::string, 14> weaponEnchantmentNames;
 		std::array<uint8_t, 14> weaponEnchantmentQualities;
 		std::array<uint8_t, 14> weaponEnchantmentSpells; // See armor enchantment note.
-		std::array<uint8_t, 14> weaponEnchantmentBonusPrices;
+		std::array<uint16_t, 14> weaponEnchantmentBonusPrices;
 
 		// Three categories of trinkets: spellcasting items, attribute enhancement items,
 		// and armor class items. Spellcasting items are split into offense/defense/misc.
 		std::array<std::string, 4> spellcastingItemNames;
 		std::array<uint8_t, 4> spellcastingItemCumulativeChances;
-		std::array<uint8_t, 4> spellcastingItemBasePrices;
+		std::array<uint16_t, 4> spellcastingItemBasePrices;
 		std::array<std::pair<uint8_t, uint8_t>, 4> spellcastingItemChargeRanges; // Min/max + 1...
 		std::array<std::string, 15> spellcastingItemAttackSpellNames;
 		std::array<uint8_t, 15> spellcastingItemAttackSpellQualities;
@@ -273,7 +273,7 @@ public:
 		std::array<uint16_t, 8> spellcastingItemMiscSpellPricesPerCharge;
 		std::array<std::string, 4> enhancementItemNames;
 		std::array<uint8_t, 4> enhancementItemCumulativeChances;
-		std::array<uint8_t, 4> enhancementItemBasePrices;
+		std::array<uint16_t, 4> enhancementItemBasePrices;
 
 		// @todo: artifacts.
